+++ packages/demo/__stories__/plugin-sc-mc-exercise.tsx
@@ -3,12 +3,22 @@ import * as React from 'react'
 
 import { EditorStory } from '../src'
 
-storiesOf('Plugins/Single-Choice-Multiple-Choice Exercise', module).add(
-  'Initial State',
-  () => {
+storiesOf('Plugins/Single-Choice-Multiple-Choice Exercise', module)
+  .add('Initial State', () => {
     const state = {
-      plugin: 'scMcExercise'
+      plugin: 'rows',
+      state: [
+        {
+          plugin: 'scMcExercise'
+        }
+      ]
     }
     return <EditorStory initialState={state} />
-  }
-)
+  })
+  .add('Prefilled Example', () => {
+    const state = JSON.parse(
+      '{"plugin":"rows","state":[{"plugin":"text","state":{"object":"value","document":{"object":"document","data":{},"nodes":[{"object":"block","type":"paragraph","data":{},"nodes":[{"object":"text","leaves":[{"object":"leaf","text":"Hier könnte Ihre Frage stehen:","marks":[]}]}]}]}}},{"plugin":"scMcExercise","state":{"isSingleChoice":false,"answers":[{"id":{"plugin":"text","state":{"object":"value","document":{"object":"document","data":{},"nodes":[{"object":"block","type":"paragraph","data":{},"nodes":[{"object":"text","leaves":[{"object":"leaf","text":"42","marks":[]}]}]}]}}},"isCorrect":true,"feedback":{"plugin":"text","state":{"document":{"nodes":[{"object":"block","type":"paragraph","nodes":[{"object":"text","leaves":[{"object":"leaf","text":""}]}]}]}}},"hasFeedback":false},{"id":{"plugin":"text","state":{"object":"value","document":{"object":"document","data":{},"nodes":[{"object":"block","type":"paragraph","data":{},"nodes":[{"object":"text","leaves":[{"object":"leaf","text":"1","marks":[]}]}]}]}}},"isCorrect":false,"feedback":{"plugin":"text","state":{"object":"value","document":{"object":"document","data":{},"nodes":[{"object":"block","type":"paragraph","data":{},"nodes":[{"object":"text","leaves":[{"object":"leaf","text":"Viel zu wenig!!","marks":[]}]}]}]}}},"hasFeedback":true},{"id":{"plugin":"text","state":{"object":"value","document":{"object":"document","data":{},"nodes":[{"object":"block","type":"paragraph","data":{},"nodes":[{"object":"text","leaves":[{"object":"leaf","text":"2","marks":[]}]}]}]}}},"isCorrect":false,"feedback":{"plugin":"text","state":{"document":{"nodes":[{"object":"block","type":"paragraph","nodes":[{"object":"text","leaves":[{"object":"leaf","text":""}]}]}]}}},"hasFeedback":false}]}}]}'
+    )
+
+    return <EditorStory defaultPlugin="text" initialState={state} />
+  })