+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to this project will be documented in this file.
 ### Internal
 
 - **demo**. Log errors only
+- **demo**. Add additional examples for plugin-usage in Storybook
 
 ## [0.4.3](https://github.com/edtr-io/edtr-io/compare/v0.4.2..v0.4.3) - March 29, 2019